requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
openai
python-dotenv
python-dateutil
